pyproject.toml
[project]
name = "video-gen"
version = "0.2.2"
description = "Claude Code skill: generate short MP4 videos from text + images. Claude directs, ffmpeg renders."
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "Pillow>=10.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
